backend/requirements.txt
# Dependencias básicas para Lambda
boto3>=1.34.0
fastapi>=0.104.1
pydantic>=2.5.0
python-dotenv>=1.0.0
mangum>=0.17.0
opensearch-py>=2.4.0
python-multipart>=0.0.6

# Dependencias adicionales
requests>=2.31.0
python-dateutil>=2.8.2
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.1.0
aiohttp>=3.8.0
